+++ .github/workflows/dotnet.yml
@@ -19,8 +19,10 @@ jobs:
     - name: Setup .NET
       uses: actions/setup-dotnet@v4
       with:
-        dotnet-version: 8.0.x
+        dotnet-version: 9.0.x
     - name: Restore dependencies
       run: dotnet restore
     - name: Build
       run: dotnet build --no-restore
+    - name: Test
+      run: dotnet test Mailjet.Tests/Mailjet.Tests.csproj --no-build --verbosity normal
