+++ Mailjet.IntegrationTests/Resources/MJMLTemplate.mjml
@@ -1,4 +1,4 @@
-﻿<mjml owa="desktop" version="4.3.0">
+<mjml owa="desktop" version="4.3.0">
 	<mj-head>
 		<mj-font href="https://fonts.googleapis.com/css?family=Montserrat" name="Montserrat"></mj-font>
 		<mj-font href="https://fonts.googleapis.com/css?family=Raleway" name="Raleway"></mj-font>
